refactor: Calculator class 파일 분리
